+++ include/socketCAN.h
@@ -23,10 +23,10 @@ int		check_mtu_support(int s, struct ifreq *ifr);
 int		socketCan_init(const char *interface);
 
 int		can_send_frame(int socket, uint16_t can_id, 
-					const int8_t *data, uint8_t len);
+					const int16_t *data, uint8_t len);
 
 int		can_send_frame_fd(int socket, uint16_t can_id, 
-					const int8_t *data, uint8_t len);
+					const int16_t *data, uint8_t len);
 
 void	can_close(int socket);
